Major refactor to remove singletons
