+++ ingest_wikimedia/web.py
@@ -3,11 +3,7 @@ import functools
 import requests
 from requests.adapters import HTTPAdapter
 from urllib3 import Retry
-import threading
 
-
-__thread_local = threading.local()
-__thread_local.http_session = None
 
 RETRY_COUNT = 3
 RETRY_BACKOUT_FACTOR = 1
@@ -16,10 +12,8 @@ DEFAULT_CONN_TIMEOUT = 45
 
 def get_http_session() -> requests.Session:
     """
-    Returns an initialized Requests session for the current thread.
+    Returns an initialized Requests session.
     """
-    if __thread_local.http_session is not None:
-        return __thread_local.http_session
     retry_strategy = Retry(
         total=RETRY_COUNT,
         backoff_factor=RETRY_BACKOUT_FACTOR,
@@ -36,7 +30,6 @@ def get_http_session() -> requests.Session:
     session.get = functools.partial(session.get, timeout=DEFAULT_CONN_TIMEOUT)
     session.mount("https://", adapter)
     session.mount("http://", adapter)
-    __thread_local.http_session = session
     return session
