+++ tools/nuke.py
@@ -8,8 +8,9 @@ from tqdm import tqdm
 
 from ingest_wikimedia.common import load_ids
 from ingest_wikimedia.logs import setup_logging
-from ingest_wikimedia.metadata import check_partner
-from ingest_wikimedia.s3 import get_item_s3_path, S3_BUCKET
+
+from ingest_wikimedia.s3 import S3_BUCKET
+from ingest_wikimedia.tools_context import ToolsContext
 
 
 @click.command()
@@ -18,13 +19,16 @@ from ingest_wikimedia.s3 import get_item_s3_path, S3_BUCKET
 @click.option("--dry-run", is_flag=True)
 def main(ids_file: IO, partner: str, dry_run: bool):
     start_time = time.time()
-    check_partner(partner)
+    tools_context = ToolsContext.init()
+    s3 = tools_context.get_s3_client()
+    dpla = tools_context.get_dpla()
+    dpla.check_partner(partner)
     setup_logging(partner, "nuke-items", logging.INFO)
     logging.info(f"Nuking items for {partner}")
     dpla_ids = load_ids(ids_file)
     for dpla_id in tqdm(dpla_ids, desc="Nuking Items", unit="Item", ncols=100):
         logging.info(f"DPLA ID: {dpla_id}")
-        s3_path = get_item_s3_path(dpla_id, "", partner)
+        s3_path = s3.get_item_s3_path(dpla_id, "", partner)
         command = f"aws s3 rm s3://{S3_BUCKET}/{s3_path} --recursive"
         if dry_run:
             command = command + " --dryrun"
